new way of estimating affine transform for tile
Instead of fitting an affine model to the offsets, average the coefficients of the linear part and average the offsets of the translational part where the linear part has been undone.
This results in more precise prediction of the tile position and orientation, and works in the case where the entire volume is rotated. The previous approach will not work in this case because the offsets will sum to a value close to 0, and this will give incorrect rotation angle.