add background intensity as parameter to flatfield estimation step
By default, the background value is estimated automatically, which in certain cases may differ from the true background value and yield poor deconvolution results in the next steps. If the background intensity is known, the user is encouraged to provide it. Then, it will be picked up by the deconvolution step.