convert preprocessed images to input data type to improve correlations
The preprocessing includes flatfield correction, averaging, and blurring and is done using Float images. If these images are then converted to the input data type (usually Short/uint16) and then used for phase-correlation and cross-correlation, it can yield more stable results than if using the Float images without conversion. This is likely due to quantizing the intensity value range, and it helps to choose the correct correlation peak and in turn increase the robustness of the global optimization and generate better solutions.