make input/output paths the same for filesystem and cloud runs
This is possible now since recent releases of n5-aws-s3 and n5-google-cloud do not require N5 containers to be buckets anymore. Now an N5 container can be stored at some specified path in a bucket just like on the filesystem, and it's possible to have multiple N5 containers per bucket as well.